+++ src/controllers/events.controllers.ts
@@ -86,6 +86,7 @@ const updateEvent: RequestHandler = async (req: any, res) => {
         await EventModel.populate(event, [
             { path: 'saves', select: 'id name email type headline bio avatar', options: { autopopulate: false } },
             { path: 'business' },
+            { path: 'thumbnail' },
         ]);
         return constructResponse({
             res,
@@ -321,9 +322,8 @@ const getEvent: RequestHandler = async (req: any, res) => {
 const getEvents: RequestHandler = async (req: any, res) => {
     try {
         const userId = req.user.id.toString()
-        // TODO HIDE BASED ON OWNER
+
         const events = await EventModel.aggregate([
-            // { $match: { saves: req.user._id } },
             { $sort: { createdAt: -1 } },
             {
                 $addFields: {
@@ -336,8 +336,8 @@ const getEvents: RequestHandler = async (req: any, res) => {
             { $unset: ["_id", "__v"] }
         ]);
         await EventModel.populate(events, [
-            // { path: 'saves', select: 'id name email type headline bio avatar', options: { autopopulate: false } },
             { path: 'business' },
+            { path: 'thumbnail' },
         ]);
         const data = paginateData(req.query, events, 'events')
 
@@ -363,7 +363,6 @@ const getMyEvents: RequestHandler = async (req: any, res) => {
     try {
         const business = req.user.business
         const userId = req.user.id.toString()
-        // TODO HIDE BASED ON OWNER
         const events = await EventModel.aggregate([
             { $match: { business: business._id } },
             { $sort: { createdAt: -1 } },
@@ -379,6 +378,7 @@ const getMyEvents: RequestHandler = async (req: any, res) => {
         await EventModel.populate(events, [
             { path: 'saves', select: 'id name email type headline bio avatar', options: { autopopulate: false } },
             { path: 'business' },
+            { path: 'thumbnail' },
         ]);
         const data = paginateData(req.query, events, 'events')
 
@@ -403,7 +403,6 @@ const getMyEvents: RequestHandler = async (req: any, res) => {
 const getSavedEvents: RequestHandler = async (req: any, res) => {
     try {
         const userId = req.user.id.toString()
-        // TODO HIDE BASED ON OWNER
         const events = await EventModel.aggregate([
             { $match: { saves: req.user._id } },
             { $sort: { createdAt: -1 } },
@@ -420,6 +419,7 @@ const getSavedEvents: RequestHandler = async (req: any, res) => {
         await EventModel.populate(events, [
             // { path: 'saves', select: 'id name email type headline bio avatar', options: { autopopulate: false } },
             { path: 'business' },
+            { path: 'thumbnail' },
         ]);
         const data = paginateData(req.query, events, 'events')
 
@@ -444,7 +444,6 @@ const getSavedEvents: RequestHandler = async (req: any, res) => {
 const getRegisteredEvents: RequestHandler = async (req: any, res) => {
     try {
         const userId = req.user.id.toString()
-        // TODO HIDE BASED ON OWNER
         const events = await EventModel.aggregate([
             { $match: { attendees: req.user._id } },
             { $sort: { createdAt: -1 } },
@@ -461,6 +460,7 @@ const getRegisteredEvents: RequestHandler = async (req: any, res) => {
         await EventModel.populate(events, [
             // { path: 'saves', select: 'id name email type headline bio avatar', options: { autopopulate: false } },
             { path: 'business' },
+            { path: 'thumbnail' },
         ]);
         const data = paginateData(req.query, events, 'events')