+++ src/controllers/users.controllers.ts
@@ -1,6 +1,6 @@
 import { RequestHandler } from 'express';
 import { Types } from 'mongoose';
-import { AccountModel, BusinessModel, ConnectionRequestModel, ProjectModel } from '../mongodb/models';
+import { AccountModel, BusinessModel, ConnectionRequestModel, ProjectModel, SubscriptionModel } from '../mongodb/models';
 import { EventsService, JobsService, constructResponse, uploadMedia } from '../services';
 import { paginateData } from '../utilities/common';
 import { AppConfig } from '../utilities/config';
@@ -62,6 +62,7 @@ const discoverUsers: RequestHandler = async (req: any, res) => {
 
         await AccountModel.populate(accounts, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
 
         const data = paginateData(req.query, accounts, 'accounts')
@@ -87,7 +88,7 @@ const discoverUsers: RequestHandler = async (req: any, res) => {
 const getAllUsers: RequestHandler = async (req: any, res) => {
     try {
         const { skills, name, connections } = req.query;
-        const userId = req.user.id.toString()
+        const userId = "65b38f6e27e1c2bb36c7d015"
         // Construct base query
         let query: any = {};
 
@@ -121,6 +122,7 @@ const getAllUsers: RequestHandler = async (req: any, res) => {
 
         await AccountModel.populate(accounts, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
 
         const data = paginateData(req.query, accounts, 'accounts')
@@ -167,10 +169,13 @@ const getUserProfile: RequestHandler = async (req: any, res) => {
 
         await AccountModel.populate(user, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
 
-        const { subscription, connections, firebaseToken, ...rest } = user.toJSON()
-        const data: any = { ...rest, connections, requested: false, connected: false, projects: [], events: [], jobs: [] }
+        const subscription = await SubscriptionModel.findOne({ account: user.id })
+
+        const { connections, firebaseToken, ...rest } = user.toJSON()
+        const data: any = { ...rest, connections, subscription, requested: false, connected: false, projects: [], events: [], jobs: [] }
         if (connectionRequests.filter(x => x.recipient.toString() === user.id.toString()).length > 0) {
             data.requested = true
         }
@@ -209,6 +214,7 @@ const getProfile: RequestHandler = async (req: any, res: any) => {
         const user = req.user
         await AccountModel.populate(user, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
         return constructResponse({
             res,
@@ -401,10 +407,13 @@ const requestConnection: RequestHandler = async (req: any, res) => {
         })
         await AccountModel.populate(recipient, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
+        
+        const subscription = await SubscriptionModel.findOne({ account: recipient.id })
 
-        const { subscription, connections, firebaseToken, ...rest } = recipient.toJSON()
-        const data = { recipient: { ...rest, connections, requested: false, connected: false }, request }
+        const { connections, firebaseToken, ...rest } = recipient.toJSON()
+        const data = { recipient: { ...rest, connections, subscription, requested: false, connected: false }, request }
         data.recipient.requested = true
         data.recipient.connected = false
         return constructResponse({
@@ -472,9 +481,10 @@ const acceptRequest: RequestHandler = async (req: any, res) => {
         await request.deleteOne()
         await AccountModel.populate(sender, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
 
-        const { subscription, connections, firebaseToken, ...rest } = sender.toJSON()
+        const { connections, firebaseToken, ...rest } = sender.toJSON()
         const data = { sender: { ...rest, connections, requested: false, connected: false }, request: null }
         data.sender.requested = false
         data.sender.connected = true
@@ -514,6 +524,7 @@ const rejectRequest: RequestHandler = async (req: any, res) => {
         const sender: any = await AccountModel.findById(request.sender)
         await AccountModel.populate(sender, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
 
         const { subscription, connections, firebaseToken, ...rest } = sender.toJSON()
@@ -558,6 +569,7 @@ const abandonRequest: RequestHandler = async (req: any, res) => {
         const sender: any = await AccountModel.findById(request.sender)
         await AccountModel.populate(sender, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
 
         const { subscription, connections, firebaseToken, ...rest } = sender.toJSON()
@@ -621,6 +633,7 @@ const removeConnection: RequestHandler = async (req: any, res) => {
 
         await AccountModel.populate(user, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
 
 
@@ -701,6 +714,7 @@ const getUserConnections: RequestHandler = async (req: any, res) => {
 
         await AccountModel.populate(user, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
         const data = paginateData(req.query, user.connections, 'connections')
         return constructResponse({
@@ -737,6 +751,7 @@ const getProfileConnections: RequestHandler = async (req: any, res) => {
 
         await AccountModel.populate(user, [
             { path: 'connections', select: 'id name email type headline bio avatar' },
+            { path: 'avatar' },
         ]);
         const data = paginateData(req.query, user.connections, 'connections')
         return constructResponse({