+++ src/docs/common.doc.ts
@@ -17,6 +17,7 @@ export const fileSwaggerSchema = {
 export const trueFileSwaggerSchema = {
     type: 'object',
     properties: {
+        _id: { type: 'string' },
         id: { type: 'string' },
         url: { type: 'string', format: 'uri' }
     }
@@ -433,7 +434,6 @@ export const jobSchema = {
         business: businessSchema,
         createdAt: { type: 'string', format: 'date-time' },
         updatedAt: { type: 'string', format: 'date-time' },
-        // TODO
         applicants: {
             anyOf: [
                 { type: 'null' },